refactor(appservice)!: Cleanup
